Improve case search validation
Refactor futureDateValidator into it's own file
Made validation summary more reactive to form changes
Fixed validation fragment links
Update basic search validation message
